+++ lib/active_interaction.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'active_model'
+require 'active_record'
 require 'active_support/core_ext/hash/indifferent_access'
 require 'action_controller'